Improved tier deployments.

+++ commands/ssh/entrypoint.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -Eeuo pipefail
+set -Eeuox pipefail
 cd "$(dirname "$0")"
 
 VALUE=${2:-}
@@ -14,7 +14,6 @@ fi
 
 SSH_USERNAME=
 SSH_HOSTNAME=
-SSH_PUSH=0
 SSH_KEY_PATH=
 ENDPOINT_DIR=
 ONION_ADDRESS=
@@ -51,10 +50,6 @@ for i in "$@"; do
             TITLE="${i#*=}"
             shift # past argument=value
         ;;
-        --push)
-            SSH_PUSH=1
-            shift # past argument=value
-        ;;
         *)
             # unknown option
         ;;
@@ -75,9 +70,9 @@ if [[ $BCM_HELP_FLAG == 1 ]]; then
     exit
 fi
 
+KEY_NAME="bcm_trezor_$SSH_HOSTNAME"".pub"
+TREZOR_PUB_KEY_PATH="$BCM_SSH_DIR/$KEY_NAME"
 if [[ $BCM_CLI_VERB == "newkey" ]]; then
-    
-    
     # shellcheck disable=SC1090
     source "$BCM_GIT_DIR/controller/export_usb_path.sh"
     if [[ -z "$BCM_TREZOR_USB_PATH" ]]; then
@@ -90,12 +85,9 @@ if [[ $BCM_CLI_VERB == "newkey" ]]; then
         exit
     fi
     
-    KEY_NAME="id_rsa_trezor.pub"
-    TREZOR_PUB_KEY_PATH="$ENDPOINT_DIR/$KEY_NAME"
-    
     docker run -t --rm \
     -v "$BCM_TREZOR_USB_PATH":"$BCM_TREZOR_USB_PATH" \
-    -v "$ENDPOINT_DIR":/home/user/.ssh \
+    -v "$BCM_SSH_DIR":/home/user/.ssh \
     -e SSH_USERNAME="$SSH_USERNAME" \
     -e SSH_HOSTNAME="$SSH_HOSTNAME" \
     -e KEY_NAME="$KEY_NAME" \
@@ -107,30 +99,24 @@ if [[ $BCM_CLI_VERB == "newkey" ]]; then
         exit
     fi
     
-    #echo "Congratulations! Your new SSH public key can be found at '$TREZOR_PUB_KEY_PATH'"
+    chmod 0400 "$TREZOR_PUB_KEY_PATH"
     
-    # We'll stop here unless instructed to push the new key.
-    if [[ $SSH_PUSH != 1 ]]; then
-        exit
-    fi
+    echo "INFO! Your new Trezor-backed SSH public key for host '$SSH_HOSTNAME' can be found at '$TREZOR_PUB_KEY_PATH'"
+fi
+
+# this command pushes a Trezor public SSH key to an existing SSH host which is AUTHENTICATED using an external SSH keypair.
+# the existing keypair will be removed from the remote node such that ONLY trezor can authenticate you to the endpoint.
+if [[ $BCM_CLI_VERB == "push" ]]; then
+    # let's add the remote hosts fingerprint to known hosts so we don't get interactive input.
+    ssh-keyscan -H "$SSH_HOSTNAME" >> "$BCM_KNOWN_HOSTS_FILE"
+    cat "$TREZOR_PUB_KEY_PATH" | ssh -i "$SSH_KEY_PATH" "$SSH_USERNAME@$SSH_HOSTNAME" "cat > /home/$SSH_USERNAME/.ssh/authorized_keys"
     
-    if [[ -f $SSH_KEY_PATH ]]; then
-        # push the trezor ssh pubkey to the destination.
-        ssh-copy-id -i "$SSH_KEY_PATH" -o UserKnownHostsFile="$BCM_KNOWN_HOSTS_FILE" "$SSH_USERNAME@$SSH_HOSTNAME"
-        
-        # remove the entry for the host in your BCM_KNOWN_HOSTS_FILE
-        ssh-keygen -f "$BCM_KNOWN_HOSTS_FILE" -R "$SSH_HOSTNAME"  >> /dev/null
-        
-        # new key is up there, now let's do a ssh-keyscan from our SDN controller
-        # so we won't get any annoying warnings about keys changing.
-        ssh-keyscan -H "$SSH_HOSTNAME" >> "$BCM_KNOWN_HOSTS_FILE"
-        
-        exit
-    fi
+    echo "WARNING: You can ONLY use your Trezor now to log into the host '$SSH_HOSTNAME'."
+    ssh-keyscan -H "$SSH_HOSTNAME" >> "$BCM_KNOWN_HOSTS_FILE"
+    exit
 fi
 
 if [[ $BCM_CLI_VERB == "connect" ]]; then
-    
     if [[ $BCM_HELP_FLAG == 1 ]]; then
         cat ./connect/help.txt
         exit
@@ -142,15 +128,59 @@ if [[ $BCM_CLI_VERB == "connect" ]]; then
         exit
     fi
     
+    if [[ -z $SSH_USERNAME ]]; then
+        echo "ERROR: SSH username not specified."
+        exit 1
+    fi
+    
     IP_ADDRESS=$(dig +short "$SSH_HOSTNAME" | head -n 1)
     docker run -it --rm --add-host="$SSH_HOSTNAME:$IP_ADDRESS" \
     -v "$BCM_TREZOR_USB_PATH":"$BCM_TREZOR_USB_PATH" \
+    -v "$BCM_SSH_DIR":/home/user/ssh \
     -e SSH_USERNAME="$SSH_USERNAME" \
     -e SSH_HOSTNAME="$SSH_HOSTNAME" \
     --device="$BCM_TREZOR_USB_PATH" \
-    "bcm-trezor:$BCM_VERSION" bash -c "trezor-agent $SSH_USERNAME@$SSH_HOSTNAME --connect --verbose"
+    "bcm-trezor:$BCM_VERSION" trezor-agent $SSH_USERNAME@$SSH_HOSTNAME --connect --verbose
     
     exit
+fi
+
+
+# prepares a remote server to use BCM.
+if [[ $BCM_CLI_VERB == "prepare" ]]; then
+    if [[ $BCM_HELP_FLAG == 1 ]]; then
+        cat ./prepare/help.txt
+        exit
+    fi
+    # shellcheck disable=SC1090
+    source "$BCM_GIT_DIR/controller/export_usb_path.sh"
+    if [[ -z "$BCM_TREZOR_USB_PATH" ]]; then
+        echo "Could not determine Trezor USB PATH."
+        exit
+    fi
+    
+    if [[ -z $SSH_USERNAME ]]; then
+        echo "ERROR: SSH username not specified."
+        exit 1
+    fi
+    
+    IP_ADDRESS=$(dig +short "$SSH_HOSTNAME" | head -n 1)
+    
+    if docker ps | grep -a ssher; then
+        docker kill ssher
+    fi
+    
+    docker system prune -f
+    
+    docker run -d \
+    --name=ssher \
+    --add-host="$SSH_HOSTNAME:$IP_ADDRESS" \
+    -v "$BCM_TREZOR_USB_PATH":"$BCM_TREZOR_USB_PATH" \
+    -v "$BCM_SSH_DIR":/home/user/ssh \
+    -e SSH_USERNAME="$SSH_USERNAME" \
+    -e SSH_HOSTNAME="$SSH_HOSTNAME" \
+    --device="$BCM_TREZOR_USB_PATH" \
+    "bcm-trezor:$BCM_VERSION" trezor-agent $SSH_USERNAME@$SSH_HOSTNAME -- wget 
 fi
 
 
@@ -231,26 +261,4 @@ if [[ $BCM_CLI_VERB == "list-onion" ]]; then
             echo "$TITLE,$ONION"
         fi
     done </etc/tor/torrc
-fi
-
-
-if [[ $BCM_CLI_VERB == "push-key" ]]; then
-    if [[ $BCM_HELP_FLAG == 1 ]]; then
-        cat ./push/help.txt
-        exit
-    fi
-    
-    # shellcheck disable=SC1090
-    source "$BCM_GIT_DIR/controller/export_usb_path.sh"
-    
-    IP_ADDRESS=$(dig +short "$SSH_HOSTNAME" | head -n 1)
-    docker run -it --rm --add-host="$SSH_HOSTNAME:$IP_ADDRESS" \
-    -v "$BCM_TREZOR_USB_PATH":"$BCM_TREZOR_USB_PATH" \
-    -v "$BCM_SSH_DIR":/home/user/.ssh \
-    -v "$GIT_REPO_DIR":/gitrepo \
-    -e GIT_CLIENT_USERNAME="$GIT_CLIENT_USERNAME" \
-    -e SSH_USERNAME="$SSH_USERNAME" \
-    -e SSH_HOSTNAME="$SSH_HOSTNAME" \
-    --device="$BCM_TREZOR_USB_PATH" \
-    "bcm-trezor:$BCM_VERSION" trezor-agent $SSH_USERNAME@$SSH_HOSTNAME -- service tor start && wait-for-it -t 10 127.0.0.1:9050 && git config --local http.proxy socks5://127.0.0.1:9050 && git config --local user.name "$GIT_CLIENT_USERNAME" && git push
 fi
